feat(jwa): removing curve 448 and open ssl since it was not being used
This will be back in future implementation, but I noticed it actually didn't support all the targets I required.
I will have to add the targets on a open ssl framework so then I can produce this work.

this commit also adds a fix: secp256k1 bug when trying to recreate uncompressed key